Accessory board version 1.2
Removed the MOSFET based level shifter for LED data and replaced
it with a TI TXS01012.  The MOSFET level shifter worked, but it
wasn't fast enough to keep good data integrity for the WS2811 LED
driver.

I also routed the relay control through the same TXS01012 level
shifter so it can be safely used to drive an external SSR instead
of a feather relay board.